made a small change on price type on my created dtos

+++ AggarApi/CORE/DTOs/Vehicle/UpdateVehicleDto.cs
@@ -24,7 +24,7 @@ namespace CORE.DTOs.Vehicle
         public VehiclePhysicalStatus PhysicalStatus { get; set; }
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public VehicleTransmission Transmission { get; set; }
-        public double PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
         public string? Requirements { get; set; }
         public string? ExtraDetails { get; set; }
         public Location Location { get; set; } = null!;
